test(accessibility): share rendered pixel support
What changed:
- extract one deterministic rendered-pixel helper for SwiftUI tests
- reuse it across accessibility adjustment and CVD simulation coverage

Why:
- remove duplicated rendering code and keep scale and color-scheme behavior consistent

Alternatives considered:
- rejected keeping private copies because they had already drifted in rendering configuration
